Updated variables

+++ Domain/src/main/java/com/sdk/domain/entity/payment/CustomerAccessItemEntity.kt
@@ -4,7 +4,7 @@ data class CustomerAccessItemEntity(
         val consumer_email: String,
         val created_at: Int,
         val customer_id: Int,
-        val expires_at: Int,
+        val expires_at: Long,
         val is_trial: Boolean,
         val item_access_id: Int,
         val item_id: Int,
@@ -16,9 +16,9 @@ data class CustomerAccessItemEntity(
         val purchased_access_fee_description: String,
         val purchased_access_fee_id: Int,
         val purchased_access_fee_type: String,
-        val purchased_amount: Int,
+        val purchased_amount: Double,
         val purchased_currency: String,
         val revoked: Int,
-        val starts_at: Int,
+        val starts_at: Long,
         val type: String
 )
